# release-20.2: geoindex,invertedidx,rowexec,encoding,...: add bounding box to invert…