--- a/pkg/sql/logictest/testdata/logic_test/distsql_stats
+++ b/pkg/sql/logictest/testdata/logic_test/distsql_stats
@@ -925,9 +925,9 @@ WHERE statistics_name = 's' AND column_names = '{geog}'
 query TIRI colnames
 SHOW HISTOGRAM $hist_id_1
 ----
-upper_bound             range_rows  distinct_range_rows  equal_rows
-'\xfd1000000000000000'  0           0                    1
-'\xfd5000000000000000'  0           0                    1
+upper_bound                                                                                 range_rows  distinct_range_rows  equal_rows
+'\x42fd1000000000000000000000000000000000bcc00000000000003ffbecde5da115a83ff661bdc396bcdc'  0           0                    1
+'\x42fd5000000000000000000000000000000000bcc00000000000003ffbecde5da115a83ff661bdc396bcdc'  0           0                    1
 
 statement ok
 DROP INDEX geo_table@geog_idx_1;
@@ -941,8 +941,8 @@ WHERE statistics_name = 's' AND column_names = '{geog}'
 query TIRI colnames
 SHOW HISTOGRAM $hist_id_1
 ----
-upper_bound             range_rows  distinct_range_rows  equal_rows
-'\xfd1000000000000000'  0           0                    1
-'\xfd4500000000000000'  0           0                    1
-'\xfd4700000000000000'  0           0                    1
-'\xfd5ad4000000000000'  0           0                    1
+upper_bound                                                                                 range_rows  distinct_range_rows  equal_rows
+'\x42fd1000000000000000000000000000000000bcc00000000000003ffbecde5da115a83ff661bdc396bcdc'  0           0                    1
+'\x42fd4500000000000000000000000000000000bcc00000000000003ffbecde5da115a83ff661bdc396bcdc'  0           0                    1
+'\x42fd4700000000000000000000000000000000bcc00000000000003ffbecde5da115a83ff661bdc396bcdc'  0           0                    1
+'\x42fd5ad4000000000000000000000000000000bcc00000000000003ffbecde5da115a83ff661bdc396bcdc'  0           0                    1