… pkg/geo/geoindex/s2_geometry_index_test.go
… pkg/geo/geoindex/testdata/s2_geography
… pkg/geo/geoindex/utils_test.go
… pkg/sql/logictest/testdata/logic_test/inverted_filter_geospatial_dist
… pkg/sql/logictest/testdata/logic_test/inverted_filter_geospatial_explain_local
… pkg/sql/logictest/testdata/logic_test/inverted_index_geospatial
… pkg/sql/opt/exec/execbuilder/testdata/geospatial
… pkg/sql/opt/exec/execbuilder/testdata/inverted_index
… pkg/sql/opt/invertedexpr/geo_expression_test.go
… pkg/sql/opt/invertedidx/geo_test.go
… pkg/sql/opt/memo/testdata/stats/inverted-geo
… pkg/sql/opt/xform/testdata/external/postgis-tutorial-idx
… pkg/sql/opt/xform/testdata/rules/limit
… pkg/sql/opt/xform/testdata/rules/select
… pkg/sql/rowexec/inverted_expr_evaluator.go
… pkg/sql/rowexec/inverted_expr_evaluator_test.go
… pkg/sql/rowexec/inverted_filterer.go
… pkg/sql/rowexec/inverted_filterer_test.go
… pkg/sql/rowexec/inverted_joiner.go
… pkg/sql/rowexec/inverted_joiner_test.go
… pkg/util/encoding/encoding.go
… pkg/util/encoding/encoding_test.go